refactor: explicit sender/receiver naming for transports
- transports.weights: WeightBroadcast -> WeightSender and
  WeightBroadcastReceiver -> WeightReceiver, with each transport file
  (filesystem, nccl, nixl) colocating its sender and receiver; the
  receiver module is gone. The engine-side byte receiver
  (inference/vllm/worker/nccl.py) keeps its name - it is the receiver's
  data plane, not a protocol party.
- transports.rollouts -> transports.batch: MicroBatchSender/Receiver ->
  BatchSender/Receiver (ZMQ/FileSystem variants included). The
  MicroBatch payload type and all list[MicroBatch] signatures are
  unchanged, as is the rollout_transport config field.
