+++ src/prime_rl/transports/weights/__init__.py
@@ -1,26 +1,47 @@
 from pathlib import Path
 
 import torch
+from httpx import AsyncClient
 
 from prime_rl.configs.trainer import LoRAConfig, WeightBroadcastConfig
 from prime_rl.trainer.parallel_dims import ParallelDims
-from prime_rl.transports.weights.base import WeightBroadcast
-from prime_rl.transports.weights.filesystem import FileSystemWeightBroadcast
-from prime_rl.transports.weights.nccl import NCCLWeightBroadcast
-from prime_rl.transports.weights.nixl import NIXLWeightBroadcast
+from prime_rl.transports.weights.base import WeightReceiver, WeightSender, prune_broadcasts_beyond
+from prime_rl.transports.weights.filesystem import FileSystemWeightReceiver, FileSystemWeightSender
+from prime_rl.transports.weights.nccl import NCCLWeightReceiver, NCCLWeightSender
+from prime_rl.transports.weights.nixl import NIXLWeightReceiver, NIXLWeightSender
+
+__all__ = [
+    "WeightReceiver",
+    "WeightSender",
+    "prune_broadcasts_beyond",
+    "setup_weight_receiver",
+    "setup_weight_sender",
+]
 
 
-def setup_weight_broadcast(
+def setup_weight_sender(
     output_dir: Path,
     config: WeightBroadcastConfig,
     parallel_dims: ParallelDims,
     lora_config: LoRAConfig | None = None,
-) -> WeightBroadcast:
+) -> WeightSender:
     if config.type == "nccl":
-        return NCCLWeightBroadcast(output_dir, config, torch.cuda.current_device())
+        return NCCLWeightSender(output_dir, config, torch.cuda.current_device())
     elif config.type == "filesystem":
-        return FileSystemWeightBroadcast(output_dir, config, lora_config)
+        return FileSystemWeightSender(output_dir, config, lora_config)
     elif config.type == "nixl":
-        return NIXLWeightBroadcast(output_dir, config, parallel_dims)
+        return NIXLWeightSender(output_dir, config, parallel_dims)
     else:
         raise ValueError(f"Invalid weight broadcast type: {config.type}")
+
+
+def setup_weight_receiver(
+    broadcast_dir: Path,
+    config: WeightBroadcastConfig,
+    admin_clients: list[AsyncClient],
+    model_name: str,
+) -> WeightReceiver:
+    receivers = {"filesystem": FileSystemWeightReceiver, "nccl": NCCLWeightReceiver, "nixl": NIXLWeightReceiver}
+    if config.type not in receivers:
+        raise ValueError(f"Invalid weight broadcast type: {config.type}")
+    return receivers[config.type](broadcast_dir, config, admin_clients, model_name)
