+++ src/prime_rl/transports/weights/nixl/__init__.py
@@ -1,5 +1,5 @@
 """NIXL weight broadcasting and transfer helpers."""
 
-from prime_rl.transports.weights.nixl.nixl import NIXLWeightBroadcast
+from prime_rl.transports.weights.nixl.nixl import NIXLWeightReceiver, NIXLWeightSender
 
-__all__ = ["NIXLWeightBroadcast"]
+__all__ = ["NIXLWeightReceiver", "NIXLWeightSender"]
