+++ tests/unit/orchestrator/test_algorithms.py
@@ -10,7 +10,7 @@ from verifiers.v1.types import AssistantMessage, ToolMessage, UserMessage
 from prime_rl.configs.algorithm import AlgoConfig, FrozenModelConfig
 from prime_rl.orchestrator.algo import EchoAlgorithm, assign_advantages, stamp_loss_routing
 from prime_rl.orchestrator.trajectories import trace_to_samples
-from prime_rl.transports.rollouts.types import TrainingSample
+from prime_rl.transports.batch.types import TrainingSample
 
 FROZEN = {"name": "org/ref-model", "base_url": "http://ref:8001/v1"}
